+++ src/mpl115a2.h
@@ -11,7 +11,7 @@ void *mpl115a2_init(int address, const char* i2c_device_filepath);
 
 void mpl115a2_close(void *_mpl);
 
-long mpl115a2_pressure(void *_mpl);
+float mpl115a2_pressure(void *_mpl);
 
 float mpl115a2_temperature(void *_mpl);
